utils.go platform imp
